poi version removed from pom, consolidated 2 classes into 1, additional revisions per last pr
